requirements.txt
fastapi
uvicorn
google-generativeai
pinecone-client
python-dotenv